Move timestamp route into `/v1/database`
This interacts better with SpacetimeDB-cloud's routing code.
