+++ crates/client-api/src/routes/mod.rs
@@ -12,7 +12,6 @@ mod internal;
 pub mod metrics;
 pub mod prometheus;
 pub mod subscribe;
-pub mod unstable;
 
 /// This API call is just designed to allow clients to determine whether or not they can
 /// establish a connection to SpacetimeDB. This API call doesn't actually do anything.
@@ -41,5 +40,4 @@ where
     axum::Router::new()
         .nest("/v1", router.layer(cors))
         .nest("/internal", internal::router())
-        .nest("/unstable", unstable::router())
 }
